Fix negative value for bigger numbers

+++ app/src/main/java/com/amr/fibonacci/helpers/Fibonacci.kt
@@ -1,14 +1,16 @@
 package com.amr.fibonacci.helpers
 
+import java.math.BigInteger
+
 class Fibonacci {
     companion object {
-        fun fib(n: Long): Long {
-            var a = 0.0.toLong()
-            var b = 1.0.toLong()
-            var c: Long
-            if (n < 0.toLong())
+        fun fib(n: Long): BigInteger {
+            var a = 0.toBigInteger()
+            var b = 1.toBigInteger()
+            var c: BigInteger
+            if (n.toBigInteger() < 0.toBigInteger())
                 throw Exception("n value should be positive value")
-            if (n == 0.toLong())
+            if (n == 0.0.toLong())
                 return a
             for (i in 2..n) {
                 c = a + b
@@ -18,4 +20,5 @@ class Fibonacci {
             return b
         }
     }
+
 }
